nvim: don't produce errors when creating view file
Fixes #42

When opening a file in neovim that doesn't already have a view session
file, a "could not read" error is produced and displayed. This commit
changes the error-catching behavior to silently catch the error for
neovim, but to allow it through for vim (which handles the error
silently already).